Rename file to contain service
