+++ src/app/app.module.ts
@@ -12,7 +12,7 @@ import {KeyboardShortcutsModule} from 'ng-keyboard-shortcuts';
 import {ModalModule} from './elitefour/modals/modal.module';
 import {ComponentsModule} from './elitefour/components/components.module';
 import {PagesModule} from './elitefour/pages/pages.module';
-import {FavoriteListsRepositoryImpl} from './elitefour/backend/favorite-list-repository-impl.service';
+import {FavoriteListsRepositoryImpl} from './elitefour/backend/favorite-lists-repository-impl.service';
 import {FavoriteListsRepository} from './elitefour/backend/favorite-lists-repository';
 
 @NgModule({
